2003-02-22 02:30 UTC+0100 Giancarlo Niccolai <gian@niccolai.ws>   * source/rtl/hbrandom.c     ! Removed an annoyng warning under BCC
  * source/vm/hvm.c
    ! Removed an annoyng warning under BCC

  * source/vm/garbage.c
  * source/vm/thread.c
    ! fixed a deadlock in GC related routines
    ! fixed GAWk flag in HB_GCALL() (could be not
      cleared upon exit)

+++ xharbour/source/vm/hvm.c
@@ -1,5 +1,5 @@
 /*
- * $Id: hvm.c,v 1.161 2003/02/19 10:16:32 jonnymind Exp $
+ * $Id: hvm.c,v 1.162 2003/02/19 12:17:50 jonnymind Exp $
  */
 
 /*
@@ -6545,7 +6545,7 @@ HB_FUNC( HB_EXECFROMARRAY )
    PHB_ITEM pArgs = NULL, pSelf = NULL, pString;
    PHB_DYNS pExecSym = NULL;
    PHB_FUNC pFunc;
-   int i;
+   ULONG i;
    ULONG ulLen, ulStart = 1;
 
    if( HB_IS_OBJECT( pFirst ) && hb_pcount() == 3)
